renamed old workdir to yuretscores
