+++ src/yuretscores/Yuret.dict
